Project setup/Dot Cockpit screen prototype/API v1
